Refactor: outsource topichandler to own component file
